fix help messages test
